# test-server-python-flask/swagger_server/models/add_voice_assistant_chat_message_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.add_voice_assistant_chat_message_response_ai_response import AddVoiceAssistantChatMessageResponseAiResponse  # noqa: F401,E501
from swagger_server.models.get_voice_assistant_chat_messages_messages import GetVoiceAssistantChatMessagesMessages  # noqa: F401,E501
from swagger_server import util


class AddVoiceAssistantChatMessageResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, user_prompt: GetVoiceAssistantChatMessagesMessages=None, ai_response: AddVoiceAssistantChatMessageResponseAiResponse=None):  # noqa: E501
        """AddVoiceAssistantChatMessageResponse - a model defined in Swagger

        :param user_prompt: The user_prompt of this AddVoiceAssistantChatMessageResponse.  # noqa: E501
        :type user_prompt: GetVoiceAssistantChatMessagesMessages
        :param ai_response: The ai_response of this AddVoiceAssistantChatMessageResponse.  # noqa: E501
        :type ai_response: AddVoiceAssistantChatMessageResponseAiResponse
        """
        self.swagger_types = {
            'user_prompt': GetVoiceAssistantChatMessagesMessages,
            'ai_response': AddVoiceAssistantChatMessageResponseAiResponse
        }

        self.attribute_map = {
            'user_prompt': 'user_prompt',
            'ai_response': 'ai_response'
        }
        self._user_prompt = user_prompt
        self._ai_response = ai_response

    @classmethod
    def from_dict(cls, dikt) -> 'AddVoiceAssistantChatMessageResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The AddVoiceAssistantChatMessageResponse of this AddVoiceAssistantChatMessageResponse.  # noqa: E501
        :rtype: AddVoiceAssistantChatMessageResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_prompt(self) -> GetVoiceAssistantChatMessagesMessages:
        """Gets the user_prompt of this AddVoiceAssistantChatMessageResponse.


        :return: The user_prompt of this AddVoiceAssistantChatMessageResponse.
        :rtype: GetVoiceAssistantChatMessagesMessages
        """
        return self._user_prompt

    @user_prompt.setter
    def user_prompt(self, user_prompt: GetVoiceAssistantChatMessagesMessages):
        """Sets the user_prompt of this AddVoiceAssistantChatMessageResponse.


        :param user_prompt: The user_prompt of this AddVoiceAssistantChatMessageResponse.
        :type user_prompt: GetVoiceAssistantChatMessagesMessages
        """

        self._user_prompt = user_prompt

    @property
    def ai_response(self) -> AddVoiceAssistantChatMessageResponseAiResponse:
        """Gets the ai_response of this AddVoiceAssistantChatMessageResponse.


        :return: The ai_response of this AddVoiceAssistantChatMessageResponse.
        :rtype: AddVoiceAssistantChatMessageResponseAiResponse
        """
        return self._ai_response

    @ai_response.setter
    def ai_response(self, ai_response: AddVoiceAssistantChatMessageResponseAiResponse):
        """Sets the ai_response of this AddVoiceAssistantChatMessageResponse.


        :param ai_response: The ai_response of this AddVoiceAssistantChatMessageResponse.
        :type ai_response: AddVoiceAssistantChatMessageResponseAiResponse
        """

        self._ai_response = ai_response