# test-server-python-flask/swagger_server/models/successful_voice_assistant_chat_creation_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class SuccessfulVoiceAssistantChatCreationResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, chat_id: str=None, title: str=None, personality_id: str=None, personality_name: str=None):  # noqa: E501
        """SuccessfulVoiceAssistantChatCreationResponse - a model defined in Swagger

        :param chat_id: The chat_id of this SuccessfulVoiceAssistantChatCreationResponse.  # noqa: E501
        :type chat_id: str
        :param title: The title of this SuccessfulVoiceAssistantChatCreationResponse.  # noqa: E501
        :type title: str
        :param personality_id: The personality_id of this SuccessfulVoiceAssistantChatCreationResponse.  # noqa: E501
        :type personality_id: str
        :param personality_name: The personality_name of this SuccessfulVoiceAssistantChatCreationResponse.  # noqa: E501
        :type personality_name: str
        """
        self.swagger_types = {
            'chat_id': str,
            'title': str,
            'personality_id': str,
            'personality_name': str
        }

        self.attribute_map = {
            'chat_id': 'chat_id',
            'title': 'title',
            'personality_id': 'personality_id',
            'personality_name': 'personality_name'
        }
        self._chat_id = chat_id
        self._title = title
        self._personality_id = personality_id
        self._personality_name = personality_name

    @classmethod
    def from_dict(cls, dikt) -> 'SuccessfulVoiceAssistantChatCreationResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The SuccessfulVoiceAssistantChatCreationResponse of this SuccessfulVoiceAssistantChatCreationResponse.  # noqa: E501
        :rtype: SuccessfulVoiceAssistantChatCreationResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def chat_id(self) -> str:
        """Gets the chat_id of this SuccessfulVoiceAssistantChatCreationResponse.


        :return: The chat_id of this SuccessfulVoiceAssistantChatCreationResponse.
        :rtype: str
        """
        return self._chat_id

    @chat_id.setter
    def chat_id(self, chat_id: str):
        """Sets the chat_id of this SuccessfulVoiceAssistantChatCreationResponse.


        :param chat_id: The chat_id of this SuccessfulVoiceAssistantChatCreationResponse.
        :type chat_id: str
        """
        if chat_id is None:
            raise ValueError("Invalid value for `chat_id`, must not be `None`")  # noqa: E501

        self._chat_id = chat_id

    @property
    def title(self) -> str:
        """Gets the title of this SuccessfulVoiceAssistantChatCreationResponse.


        :return: The title of this SuccessfulVoiceAssistantChatCreationResponse.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this SuccessfulVoiceAssistantChatCreationResponse.


        :param title: The title of this SuccessfulVoiceAssistantChatCreationResponse.
        :type title: str
        """
        if title is None:
            raise ValueError("Invalid value for `title`, must not be `None`")  # noqa: E501

        self._title = title

    @property
    def personality_id(self) -> str:
        """Gets the personality_id of this SuccessfulVoiceAssistantChatCreationResponse.


        :return: The personality_id of this SuccessfulVoiceAssistantChatCreationResponse.
        :rtype: str
        """
        return self._personality_id

    @personality_id.setter
    def personality_id(self, personality_id: str):
        """Sets the personality_id of this SuccessfulVoiceAssistantChatCreationResponse.


        :param personality_id: The personality_id of this SuccessfulVoiceAssistantChatCreationResponse.
        :type personality_id: str
        """
        if personality_id is None:
            raise ValueError("Invalid value for `personality_id`, must not be `None`")  # noqa: E501

        self._personality_id = personality_id

    @property
    def personality_name(self) -> str:
        """Gets the personality_name of this SuccessfulVoiceAssistantChatCreationResponse.


        :return: The personality_name of this SuccessfulVoiceAssistantChatCreationResponse.
        :rtype: str
        """
        return self._personality_name

    @personality_name.setter
    def personality_name(self, personality_name: str):
        """Sets the personality_name of this SuccessfulVoiceAssistantChatCreationResponse.


        :param personality_name: The personality_name of this SuccessfulVoiceAssistantChatCreationResponse.
        :type personality_name: str
        """
        if personality_name is None:
            raise ValueError("Invalid value for `personality_name`, must not be `None`")  # noqa: E501

        self._personality_name = personality_name