# test-server-python-flask/swagger_server/models/update_voice_assistant_chat.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class UpdateVoiceAssistantChat(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, title: str=None, personality: str=None):  # noqa: E501
        """UpdateVoiceAssistantChat - a model defined in Swagger

        :param title: The title of this UpdateVoiceAssistantChat.  # noqa: E501
        :type title: str
        :param personality: The personality of this UpdateVoiceAssistantChat.  # noqa: E501
        :type personality: str
        """
        self.swagger_types = {
            'title': str,
            'personality': str
        }

        self.attribute_map = {
            'title': 'title',
            'personality': 'personality'
        }
        self._title = title
        self._personality = personality

    @classmethod
    def from_dict(cls, dikt) -> 'UpdateVoiceAssistantChat':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UpdateVoiceAssistantChat of this UpdateVoiceAssistantChat.  # noqa: E501
        :rtype: UpdateVoiceAssistantChat
        """
        return util.deserialize_model(dikt, cls)

    @property
    def title(self) -> str:
        """Gets the title of this UpdateVoiceAssistantChat.


        :return: The title of this UpdateVoiceAssistantChat.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this UpdateVoiceAssistantChat.


        :param title: The title of this UpdateVoiceAssistantChat.
        :type title: str
        """
        if title is None:
            raise ValueError("Invalid value for `title`, must not be `None`")  # noqa: E501

        self._title = title

    @property
    def personality(self) -> str:
        """Gets the personality of this UpdateVoiceAssistantChat.


        :return: The personality of this UpdateVoiceAssistantChat.
        :rtype: str
        """
        return self._personality

    @personality.setter
    def personality(self, personality: str):
        """Sets the personality of this UpdateVoiceAssistantChat.


        :param personality: The personality of this UpdateVoiceAssistantChat.
        :type personality: str
        """
        if personality is None:
            raise ValueError("Invalid value for `personality`, must not be `None`")  # noqa: E501

        self._personality = personality